task-1/requirements.txt
# requirements.txt
flask==2.0.1
Werkzeug==2.2.2
Flask-PyMongo